+++ jdaviz/configs/imviz/plugins/viewers.py
@@ -25,7 +25,8 @@ class ImvizImageView(JdavizViewerMixin, BqplotImageView, AstrowidgetsImageViewer
                     ['jdaviz:boxzoommatch', 'jdaviz:boxzoom'],
                     ['jdaviz:panzoommatch', 'jdaviz:imagepanzoom'],
                     ['bqplot:circle', 'bqplot:rectangle', 'bqplot:ellipse',
-                     'jdaviz:singlepixelregion', 'jdaviz:clicktomovespatialregion'],
+                     'jdaviz:singlepixelregion', 'jdaviz:clicktomovespatialregion',
+                     'jdaviz:clicktocreatecircularannulus'],
                     ['jdaviz:blinkonce', 'jdaviz:contrastbias'],
                     ['jdaviz:sidebar_plot', 'jdaviz:sidebar_export', 'jdaviz:sidebar_compass']
                 ]